requirements.txt
Django==2.2.2
factory-boy==2.12.0
